[refact]: add hit_object directory
