+++ example.rt
@@ -1,8 +1,8 @@
 A 0.2 1,1,1
-C  -5,0,20 0,0,-1 70
+C -5,0,20 0,0,-1 70
 # L -15,15,0 0.3 1,0,0
 # L -25,50,-20 0.8 1,1,1
-L 0,10,-20 1 1,1,1
+L -5,10,-20 1 1,1,1
 # sp 0,10,-20 7 0,0,0.5
 # sp 4,4,-5 4 0,0.5,0
 # sp 3,0,-5 4 0,0.5,0.5
