+++ print_scene.c
@@ -1,51 +1,15 @@
 #include "miniRT.h"
 
+#define ka 0.1
+
 t_vec	set_face_normal(t_vec normal, t_ray ray)
 {
 	if (vec_dot(ray.dir, normal) > 0.0)
 	{
-		return (vec(-normal.x, -normal.y, -normal.z));  // 내각한 결과가 양수이면 카메라가 구의 안쪽에 있다는 뜻. 법선벡터의 방향 바꿔주기
+        // 내각한 결과가 양수이면 카메라가 구의 안쪽에 있다는 뜻. 법선벡터의 방향 바꿔주기
+		return (vec(-normal.x, -normal.y, -normal.z));
 	}
 	return (normal);  // 카메라가 구의 바깥에 있으면 그대로 리턴하기
-}
-
-void    set_rec(t_hit_record *rec, t_sphere *sp, t_ray ray, double root)
-{
-    t_vec   normal;
-
-    rec->albedo = sp->albedo;
-	rec->t = root;
-    rec->tmax = rec->t;
-	rec->p = ray_at(ray, rec->t);
-	// 법선 벡터의 방향을 계산해서 써줘야 함.
-	normal = vec_div((vec_sub(rec->p, sp->center)), sp->radius);
-    // normal = vec_unit(vec_sub(rec->p, sp->center));
-	rec->normal = set_face_normal(normal, ray);  // 안, 밖 고려해서 법선벡터 바꾸기
-}
-
-int hit_sphere(t_hit_record *rec, t_ray ray, t_sphere *sp)
-{
-    double  root;
-    double  discriminant;
-    t_util  util;
-
-    util.oc = vec_sub(ray.origin, sp->center);
-    util.a = vec_dot(ray.dir, ray.dir);
-    util.b = 2.0 * vec_dot(util.oc, ray.dir);
-    util.c = vec_dot(util.oc, util.oc) - (sp->radius * sp->radius);
-    discriminant = (util.b * util.b) - 4 * (util.a * util.c);
-	if (discriminant < 0)  // 판별식이 음수이면 해가 없음.
-		return (0);
-    util.sqrtd = sqrt(discriminant);
-    root = (-util.b - util.sqrtd) / (2 * util.a);
-	if (root < rec->tmin || root > rec->tmax)
-	{
-		root = (-util.b + util.sqrtd) / (2 * util.a);
-		if (root < rec->tmin || root > rec->tmax)
-			return (0);
-	}
-    set_rec(rec, sp, ray, root);
-	return (1);
 }
 
 int rgb_to_color(t_vec color)
@@ -64,7 +28,7 @@ int ray_color(t_ray ray, t_scene scene)
     color = vec(1, 1, 1);    // color값 초기화
     scene.rec.tmin = 0;
     scene.rec.tmax = INFINITY;
-    scene.ambient.color = vec_mul(scene.ambient.color, 0.1);   // define ka : 0.1
+    scene.ambient.color = vec_mul(scene.ambient.color, ka);
     obj_list = scene.objects;
     while (obj_list)
     {
@@ -76,9 +40,7 @@ int ray_color(t_ray ray, t_scene scene)
         else if (obj_list->type == CYLINDER)
             is_hit = hit_cylinder(&(scene.rec), ray, obj_list->element);
         if (is_hit == 1)
-        {
             color = phong_light(&scene, obj_list->id);
-        }
         obj_list = obj_list->next;
     }
     return (rgb_to_color(color));
@@ -115,7 +77,6 @@ void print_scene(t_scene scene, t_data image)
 
     // camera view 추가
     t_vec   lookfrom = scene.camera.origin;
-    // t_vec   lookat = vec_unit(vec_add(lookfrom, scene.camera.dir));
     t_vec   lookat = vec_add(lookfrom, vec_unit(scene.camera.dir));
     t_vec   vup = vec(0, 1, 0);
     scene.viewport.focal_length = vec_length(vec_sub(lookfrom, lookat));
