+++ shadow.c
@@ -16,10 +16,11 @@ int	hit(t_scene *scene, t_ray r, t_hit_record *rec, int id)
 {
 	int				hit_anything;
 	t_hit_record	tmp_rec;
+	t_object		*iter;
 
-	tmp_rec = *rec;
 	hit_anything = 0;
-	t_object *iter = scene->objects;
+	tmp_rec = *rec;
+	iter = scene->objects;
 	while (iter)
 	{
 		if (iter->id != id)
@@ -31,7 +32,6 @@ int	hit(t_scene *scene, t_ray r, t_hit_record *rec, int id)
 				*rec = tmp_rec;
 			}
 			// 평면 추가하기
-			// 원기둥 추가하기
 			if (hit_cylinder(rec, r, iter->element))
 			{
 				hit_anything = 1;
@@ -52,6 +52,5 @@ int	in_shadow(t_scene *scene, t_ray r, double light_len, int id)
 	rec.tmax = light_len;
 	if (hit(scene, r, &rec, id))
 		return (1);
-		return (0);
 	return (0);
 }
